refactor: share wine utils
